backend/tests/test_admin_2fa_flow.py
from datetime import datetime, timedelta

import pytest

import appy


@pytest.fixture
def enable_email_2fa(monkeypatch):
    sent_codes = []

    monkeypatch.setattr(appy, "EMAIL_2FA_ENABLED", True)
    # Also set the Flask app config to ensure the running app observes the flag
    try:
        if hasattr(appy, "app") and getattr(appy, "app") is not None:
            appy.app.config["EMAIL_2FA_ENABLED"] = True
    except Exception:
        pass

    code_sequence = iter(["123456", "654321", "789012", "210987"])

    def _next_code():
        return next(code_sequence)

    monkeypatch.setattr(appy, "generate_email_2fa_code", _next_code)

    def _capture(code, ip_address, user_agent):
        sent_codes.append(code)
        return True

    monkeypatch.setattr(appy, "send_email_2fa_code", _capture)
    return sent_codes


def _login_with_credentials(client):
    response = client.post(
        "/admin/login",
        data={"username": "admin", "password": "Admin123"},
        follow_redirects=True,
    )
    return response


@pytest.fixture
def admin_login(client, enable_email_2fa, db_session, set_pending_admin_session):
    # Ensure admin user exists in the same app context / DB session the client will use.
    try:
        from appy import db
        try:
            # Prefer canonical models import
            from models import AdminUser, User
        except Exception:
            from helpchain_backend.src.models import AdminUser, User  # type: ignore
        with client.application.app_context():
            try:
                if db.session.query(AdminUser).filter_by(username="admin").count() == 0:
                    # create minimal admin user matching tests expectations
                    admin = AdminUser(username="admin", email="admin@helpchain.live")
                    # set password via whatever helper exists; fallback to plaintext field
                    try:
                        admin.set_password("Admin123")
                    except Exception:
                        # if set_password not available, set a compatible hash directly
                        from werkzeug.security import generate_password_hash
                        admin.password_hash = generate_password_hash("Admin123")
                    db.session.add(admin)
                    db.session.commit()
            except Exception:
                # If seeding fails, continue; login attempt will reveal the issue
                pass
    except Exception:
        pass
    response = _login_with_credentials(client)
    assert response.status_code == 200
    # Some test runs may not reach the redirect reliably due to timing; ensure
    # the test client session is in the pending 2FA state so resend/success
    # flow can be exercised deterministically.
    try:
        code = None
        if enable_email_2fa:
            # enable_email_2fa returns the sent_codes list; if it's already
            # populated take the first value, otherwise fall back to default.
            code = enable_email_2fa[0] if len(enable_email_2fa) else None
    except Exception:
        code = None
    if not code:
        code = "123456"
    # Ensure pending_admin_id is present so resend handler won't bail out.
    admin_id = None
    try:
        from appy import db
        try:
            from models import AdminUser
        except Exception:
            from helpchain_backend.src.models import AdminUser  # type: ignore
        admin_obj = db.session.query(AdminUser).filter_by(username="admin").first()
        if admin_obj:
            admin_id = getattr(admin_obj, "id", None)
    except Exception:
        admin_id = None

    # Use central helper to set the pending admin session keys
    result = set_pending_admin_session(client, code=code, expires_seconds=300)
    # Fast-fail if helper could not set an admin id — clearer error message than a redirect
    assert result.get("admin_id") is not None, "Failed to set pending_admin_id via helper"
    return response


def test_email_2fa_invalid_code(client, admin_login):
    del admin_login  # ensure fixture is exercised

    with client.session_transaction() as session:
        real_code = session.get("email_2fa_code")
    assert real_code, "Очаквахме генериран 2FA код"

    invalid_response = client.post(
        "/admin/email_2fa",
        data={"code": "000000"},
        follow_redirects=True,
    )
    assert invalid_response.status_code == 200
    html = invalid_response.get_data(as_text=True).lower()
    assert "/admin/email_2fa" in (invalid_response.request.path or "")
    assert "невалиден" in html or "грешен" in html


def test_email_2fa_expired_code(client, admin_login):
    del admin_login

    with client.session_transaction() as session:
        real_code = session.get("email_2fa_code")
        assert real_code, "2FA кодът трябва да е наличен преди изтичане"
        session["email_2fa_expires"] = 0  # насилваме изтичане

    expired_response = client.post(
        "/admin/email_2fa",
        data={"code": real_code},
        follow_redirects=True,
    )
    assert expired_response.status_code == 200
    html = expired_response.get_data(as_text=True).lower()
    assert "изтек" in html or "валидност" in html or "опитайте отново" in html
    assert "/admin/login" in (expired_response.request.path or "")


def test_email_2fa_resend_and_success(client, admin_login):
    del admin_login

    with client.session_transaction() as session:
        original_code = session.get("email_2fa_code")
    assert original_code, "Очаквахме първоначален 2FA код"

    resend_response = client.get("/admin/email_2fa/resend", follow_redirects=True)
    assert resend_response.status_code == 200
    resend_html = resend_response.get_data(as_text=True).lower()
    assert "нов код" in resend_html or "изпратен" in resend_html

    with client.session_transaction() as session:
        refreshed_code = session.get("email_2fa_code")
    assert refreshed_code, "Трябва да има нов код след resend"
    assert refreshed_code != original_code, "Resend трябва да промени кода"

    success_response = client.post(
        "/admin/email_2fa",
        data={"code": refreshed_code},
        follow_redirects=True,
    )
    assert success_response.status_code == 200
    success_html = success_response.get_data(as_text=True).lower()
    assert (
        "dashboard" in success_html
        or "админ" in success_html
        or "табло" in success_html
    )

    with client.session_transaction() as session:
        assert session.get("admin_logged_in") is True
        assert session.get("pending_email_2fa") is None
        assert session.get("email_2fa_code") is None


def test_email_2fa_template_renders(client, db_session):
    with client.session_transaction() as session:
        session["pending_email_2fa"] = True
        session["email_2fa_code"] = "111111"
        session["email_2fa_expires"] = (
            datetime.now() + timedelta(minutes=5)
        ).timestamp()

    response = client.get("/admin/email_2fa")
    assert response.status_code == 200
    html = response.get_data(as_text=True).lower()
    assert "двуфактор" in html or "верификац" in html
