requirements.txt
flask
gunicorn
flask_sqlalchemy
flask_mail
flask_babel
flask_wtf
email_validator
flask-admin
flask-login
flask-talisman
flask-limiter
flask-cors
flask-jwt-extended
python-dotenv
flask-migrate
pyotp
flask-socketio
python-socketio
aiortc
opencv-python
asgiref
sentry-sdk[flask]
celery
redis
psycopg2-binary
# ML dependencies for predictive analytics
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
# Security updates for vulnerable dependencies
urllib3>=2.5.0
requests>=2.32.4
Pillow>=11.3.0
gradio>=5.35.0
starlette>=0.47.2
transformers>=4.53.0
regex>=2025.2.10
